Closes #8, refactor maded.
