+++ app/Http/Controllers/Api/Groups/UpdateController.php
@@ -4,7 +4,7 @@ namespace App\Http\Controllers\Api\Groups;
 
 use App\Group;
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Api\Groups\UpdateRequest;
+use App\Http\Requests\Api\Models\UpdateRequest;
 use App\Repositories\FillRepository;
 use App\Repositories\PhotoRepository;
 use Illuminate\Http\JsonResponse;
