+++ app/Http/Controllers/Api/Users/UpdateController.php
@@ -3,7 +3,7 @@
 namespace App\Http\Controllers\Api\Users;
 
 use App\Http\Controllers\Controller;
-use App\Http\Requests\Api\Users\UpdateRequest;
+use App\Http\Requests\Api\Models\UpdateRequest;
 use App\Repositories\FillRepository;
 use App\Repositories\PhotoRepository;
 use App\User;
