--- app/Http/Requests/Api/Pets/UpdateRequest.php
@@ -1,36 +1,0 @@
-<?php
-
-namespace App\Http\Requests\Api\Pets;
-
-use App\Rules\Coordinate;
-use Illuminate\Foundation\Http\FormRequest;
-use Illuminate\Validation\Rule;
-
-class UpdateRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     *
-     * @return bool
-     */
-    public function authorize()
-    {
-        return $this->user('api')->can('update', $this->route('pet'));
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array
-     */
-    public function rules()
-    {
-        return [
-            'coordinates' => [
-                Rule::requiredIf(function () {
-                    return $this->has('photo');
-                }), new Coordinate()
-            ],
-        ];
-    }
-}
